requirements.txt
PyYAML
pyroute2
.
